+++ Car_rental_PAI_project/src/main/java/com/example/Car_rental_PAI_project/service/RentalOfficeService.java
@@ -5,7 +5,7 @@ import com.example.Car_rental_PAI_project.model.RentalOffice;
 import java.util.Optional;
 
 public interface RentalOfficeService {
-    Optional<RentalOffice> getRentalOffice(Long office_Id);
+    Optional<RentalOffice> getRentalOffice(Integer office_Id);
     RentalOffice setRentalOffice(RentalOffice rentalOffice);
-    void deleteRentalOffice(Long office_Id);
+    void deleteRentalOffice(Integer office_Id);
 }
