+++ Car_rental_PAI_project/src/main/java/com/example/Car_rental_PAI_project/service/WorkerService.java
@@ -5,7 +5,7 @@ import com.example.Car_rental_PAI_project.model.Worker;
 import java.util.Optional;
 
 public interface WorkerService {
-    Optional<Worker> getWorker(Long worker_Id);
+    Optional<Worker> getWorker(Integer worker_Id);
     Worker setWorker(Worker worker);
-    void deleteWorker(Long worker_Id);
+    void deleteWorker(Integer worker_Id);
 }
